Update custom tools doc
